fix(chassis): move the exception hook to the right place
